move pg code to postgres folder + update names and paths
