+++ sdc/db/postgres/to_json_query.sql
